+++ CCL.Types/Proxies/Weather/DecalLayerProxy.cs
@@ -1,0 +1,31 @@
+﻿using UnityEngine;
+
+namespace CCL.Types.Proxies.Weather
+{
+    public class DecalLayerProxy
+    {
+        private DecalLayerChannelProxy _channel2;
+        private DecalLayerChannelProxy _channel3;
+        private DecalLayerChannelProxy _channel4;
+        private DecalLayerChannelProxy _channel1;
+        private Texture2D _layerMask;
+        private Vector2 _layerMaskOffset;
+        private Vector2 _layerMaskScale;
+        private bool _editorSectionFoldout;
+
+        public DecalLayerProxy()
+        {
+            _layerMaskScale = new Vector2(1f, 1f);
+            _layerMaskOffset = new Vector2(0f, 0f);
+
+            _channel1 = new DecalLayerChannelProxy(DecalLayerChannelProxy.DecalChannelMode.SimpleRangeRemap);
+
+            _channel2 = new DecalLayerChannelProxy();
+            _channel3 = new DecalLayerChannelProxy();
+            _channel4 = new DecalLayerChannelProxy();
+
+            _layerMask = null!;
+            _editorSectionFoldout = false;
+        }
+    }
+}
