+++ CCL.Types/Proxies/Weather/WetDecalEnums.cs
@@ -1,0 +1,31 @@
+﻿namespace CCL.Types.Proxies.Weather
+{
+    public class WetDecalEnums
+    {
+        public enum LayerMode
+        {
+            None,
+            Single,
+            Triplanar
+        }
+
+        public enum ProjectionMode
+        {
+            Local,
+            World
+        }
+
+        public enum WetDecalMode
+        {
+            Wet,
+            Dry
+        }
+
+        public enum DecalShape
+        {
+            Cube,
+            Sphere,
+            Mesh
+        }
+    }
+}
