+++ CCL.Types/Proxies/Weather/WindowProxy.cs
@@ -1,11 +1,7 @@
-﻿using System;
-using System.Collections.Generic;
-using System.Linq;
-using System.Text;
-using System.Threading.Tasks;
+﻿using System.Linq;
 using UnityEngine;
 
-namespace CCL.Types.Proxies
+namespace CCL.Types.Proxies.Weather
 {
     [DisallowMultipleComponent]
     public class WindowProxy : MonoBehaviour
@@ -24,7 +20,7 @@ namespace CCL.Types.Proxies
         public Rigidbody rb;
 
         [RenderMethodButtons]
-        [MethodButton("CCL.Types.Proxies.WindowProxy:SetupDuplicates", "Setup duplicates")]
+        [MethodButton("CCL.Types.Proxies.Weather.WindowProxy:SetupDuplicates", "Setup duplicates")]
         public bool buttonRender;
 
         private void OnDrawGizmosSelected()
